+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.5.1
+
+- Respect languageIds in `scope` for VSCode snippets.
+- Respect languageId in comment for VSCode snippets.
+- Load global scoped VSCode snippets for all filetypes.
+
 # 2.5.0
 
 - Load snippets when necessary.